+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.1.5] - 2025-10-22
+
+### Fixed
+- Display package version as "v1.2.3" instead of "@1.2.3" in hover tooltips to prevent mailto link rendering.
+
 ## [0.1.4] - 2025-09-30
 
 ### Added